LayoutTests/webgpu/render-pipelines.html

<!DOCTYPE html>
<html>
<script src="../resources/js-test-pre.js"></script>
<script src="js/set-up-webgpu-contexts.js"></script>
<script id="library" type="x-shader/x-metal">
    #include <metal_stdlib>
    
    using namespace metal;
    
    struct Vertex
    {
        float4 position [[position]];
    };
    
    vertex Vertex vertex_main(uint vid [[vertex_id]])
    {
        Vertex v;
        switch (vid) {
        case 0:
            v.position = float4(-.75, -.75, 0, 1);
            break;
        case 1:
            v.position = float4(.75, -.75, 0, 1);
            break;
        case 2:
            v.position = float4(0, .75, 0, 1);
            break;
        default:
            v.position = float4(0, 0, 0, 1);
        }
        return v;
    }
    
    fragment float4 fragment_main(Vertex vertexIn [[stage_in]])
    {
        return float4(1.0, 0.0, 0.0, 1.0);
    }
</script>
<script>
'use strict';

if (window.testRunner)
    window.testRunner.dumpAsText();

let shaderDescriptor, shaderModule, vertexStageDescriptor, fragmentStageDescriptor;

function setUpPipeline() {
    shaderDescriptor = {
        code: document.getElementById("library").text
    };

    shaderModule = defaultDevice.createShaderModule(shaderDescriptor);
    if (!shaderModule) {
        testFailed("Could not create WebGPUShaderModule!");
        return;
    }

    vertexStageDescriptor = { 
        module: shaderModule, 
        stage: WebGPUShaderStage.VERTEX, 
        entryPoint: "vertex_main" 
    };

    fragmentStageDescriptor = {
        module: shaderModule,
        stage: WebGPUShaderStage.FRAGMENT,
        entryPoint: "fragment_main"
    };

    let pipelineDescriptor = {
        stages: [vertexStageDescriptor, fragmentStageDescriptor],
        primitiveTopology: "triangleList"
    };

    let renderPipeline = defaultDevice.createRenderPipeline(pipelineDescriptor);
    if (renderPipeline)
        testPassed("Successfully created WebGPURenderPipeline.")
    else
        testFailed("Could not create WebGPURenderPipeline!");
}

function checkBadRenderPipeline(descriptor, testSubjectName) {
    let pipeline = defaultDevice.createRenderPipeline(descriptor);

    if (pipeline)
        testFailed(`WebGPURenderPipeline was created with an invalid ${testSubjectName}!`);
    else
        testPassed(`WebGPURenderPipeline with invalid ${testSubjectName} was not created.`);
}

function setUpBadPipelines() {
    checkBadRenderPipeline({}, "WebGPURenderPipelineDescriptor");

    let noModuleVertexDescriptor = {
        stage: WebGPUShaderStage.VERTEX,
        entryPoint: "vertex_main"
    }
    let noModulePipelineDescriptor = {
        stages: [noModuleVertexDescriptor, fragmentStageDescriptor],
        primitiveTopology: "triangleList"
    }
    checkBadRenderPipeline(noModulePipelineDescriptor, "shader module");

    // A (Metal) renderpipeline must have a vertex function.
    let noVertexPipelineDescriptor = {
        stages: [fragmentStageDescriptor],
        primitiveTopology: "triangleList"
    }
    checkBadRenderPipeline(noVertexPipelineDescriptor, "vertex shader stage")

    let badEntryPointDescriptor = { 
        module: shaderModule, 
        stage: WebGPUShaderStage.VERTEX, 
        entryPoint: "Vertex_Main" 
    };
    let badEntryPointPipelineDescsriptor = {
        stages: [badEntryPointDescriptor, fragmentStageDescriptor],
        primitiveTopology: "triangleList"
    }
    checkBadRenderPipeline(badEntryPointPipelineDescsriptor, "vertex shader entry point");

    let badStageEnumDescriptor = { 
        module: shaderModule, 
        stage: WebGPUShaderStage.COMPUTE, 
        entryPoint: "vertex_main" 
    };
    let badStageEnumPipelineDescriptor = {
        stages: [badStageEnumDescriptor, fragmentStageDescriptor],
        primitiveTopology: "triangleList"
    }
    checkBadRenderPipeline(badStageEnumPipelineDescriptor, "WebGPUShaderStageEnum");
}

runWebGPUTests([setUpPipeline, setUpBadPipelines]);

successfullyParsed = true;
</script>
<script src="../resources/js-test-post.js"></script>
</html>
